requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
bootstrap4==0.1.0
certifi==2024.7.4
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.0.7
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-drf-filepond==0.5.0
django-extensions==3.2.3
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
fontawesomefree==6.6.0
gunicorn==22.0.0
idna==3.7
joblib==1.4.2
mysqlclient==2.2.4
notifications==0.3.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
